+++ cattleia-backend/.env.example
@@ -10,4 +10,4 @@ INFURA_RINKEBY=
 GANACHE=
 STRIPE_SECRET=
 STRIPE_PUBLISHABLE=
-TEST_PRIVATE_KEY
+TEST_PRIVATE_KEY=
