+++ cattleia-backend/.gitignore
@@ -1,4 +1,6 @@
 node_modules/
+coverage/
 dist/
+junit.xml
 .env
 .secret
