+++ cattleia-backend/package.json
@@ -37,12 +37,14 @@
     "@types/morgan": "^1.9.3",
     "@types/supertest": "^2.0.11",
     "jest": "^27.3.1",
+    "jest-junit": "^13.0.0",
     "morgan": "^1.10.0",
     "supertest": "^6.1.6",
     "superwstest": "^1.8.0",
     "ts-jest": "^27.0.7",
     "ts-node-dev": "^1.1.8",
-    "typescript": "^4.3.5"
+    "typescript": "^4.3.5",
+    "wsreq": "^1.1.2"
   },
   "dependencies": {
     "argon2": "^0.28.2",
