+++ cattleia-backend/src/app.ts
@@ -1,8 +1,8 @@
 import cookieParser from "cookie-parser";
-import WebSocket from "./websocket/core";
+import WebSocket from "./websocket";
 import express from "express";
 import morgan from "morgan";
-import events from "./websocket";
+import events from "./events";
 import config from "./config";
 import cors from "cors";
 import http from "http";
