+++ cattleia-backend/src/events/index.ts
@@ -1,0 +1,10 @@
+import { EventStack } from "../websocket";
+
+const stack = new EventStack();
+
+stack.push("ping", function (data) {
+  console.log(data);
+  this.emit("pong", { msg: "test" });
+});
+
+export default stack;
